Rename server-add to server-modal-form
Since the same component is used to add and edit. Makes more sense change the name.
